Cleaned up the mess and the duplicated things
