+++ src/main/kotlin/presets.kt
@@ -1,5 +1,6 @@
-// This file is based on hyfetch, and was last updated 2024-06-12
+// This file adds presets. PridePresets are based on hyfetch, and were last updated 2024-06-12
 // https://github.com/hykilpikonna/hyfetch/blob/7534371b05ee877cd3c4c3733b13d7c41d09c3e/hyfetch/presets.py
+
 
 package chattore
 
@@ -519,6 +520,56 @@ val prideColors = mapOf(
     ),
 )
 
-val pridePresets = prideColors.mapValues { (_, colors) ->
-        "<gradient:${colors.joinToString(':'.toString())}><username></gradient>"
-    }.toSortedMap()
+val birdflopColors = mapOf(
+    "birdflop" to arrayOf(
+        "#084CFB",
+        "#ADF3FD"
+    ),
+    "SimplyMC" to arrrayOf (
+        "#084CFB",
+        "#ADF3FD"
+    ),
+    "Rainbow" to arrrayOf (
+        "#FF0000",
+        "#FF7F00",
+        "#FFFF00",
+        "#00FF00",
+        "#0000FF",
+        "#EAE4AA"
+    ),
+    "Skyline" to arrrayOf (
+        "#1488CC",
+        "#2B32B2"
+    ),
+    "Mango" to arrrayOf (
+        "#1488CC",
+        "#2B32B2"
+    ),
+    "Vice City" to arrrayOf (
+        "#1488CC",
+        "#2B32B2"
+    ),
+    "Dawn" to arrrayOf (
+        "#1488CC",
+        "#2B32B2"
+    ),
+    "Rose" to arrrayOf (
+        "#1488CC",
+        "#2B32B2"
+    ),
+    "Firewatch" to arrrayOf (
+        "#1488CC",
+        "#2B32B2"
+    ),        
+)
+
+val playerColors = mapOf(
+    "Waffle" to arrrayOf (
+        "#1488CC",
+        "#2B32B2"
+    ),
+)
+
+val pridePresets = prideColors.toGradientMinimessage()
+val birdflopPresets = birdflopColors.toGradientMinimessage()
+val playerPresets = playerColors.toGradientMinimessage()
